Mac 과 같은 unix 환경에서 파일 중복 시 그대로 엎어쓰는 문제가 있던 점을 회피하기 위하여 중복을 미리 판정하는 로직을 추가함.

+++ convert-filename-epoch/converter.py
@@ -39,8 +39,12 @@ def convert1970todate(filename):
     datetime_filestr = datetime_name.strftime("%Y%m%d-%H%M%S") + extstr
     
     basepath = os.getcwd()
-    os.rename(filename, datetime_filestr)
-    print("file.find: [" + filename + "] to [" + datetime_filestr + "]")
+    if os.path.isfile(datetime_filestr):
+        print("failed to rename: [" + datetime_filestr + "] already exists.")
+    else:
+        os.rename(filename, datetime_filestr)
+        print("file.find: [" + filename + "] to [" + datetime_filestr + "]")
+
     return
 
 while True:
